src/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
dj-database-url==0.5.0
Django==1.11.13
django-ajax==0.2.0
django-ajax-comments==1.1.0.dev1
django-allauth==0.36.0
django-ckeditor==5.3.0
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-el-pagination==3.2.4
django-extensions==2.1.1
django-froala-editor==2.8.4
django-hitcount==1.3.0
django-js-asset==0.1.1
django-likes==2.0.0
django-material==1.0.0
django-modelcluster==3.1
django-secretballot==1.0.0
django-storages==1.6.6
django-taggit==0.22.1
django-taggit-labels==0.5.1
django-threadedcomments==1.2
django-treebeard==4.1.2
djangorestframework==3.6.3
dropbox==9.0.0
enum34==1.1.6
gunicorn==19.9.0
idna==2.6
ipaddress==1.0.22
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
olefile==0.45.1
Pillow==4.2.1
psycopg2==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
python-openid==2.2.5
pytz==2018.5
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sorl-thumbnail==12.3
typing==3.6.6
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.14.1
whitenoise==4.0
